Solved compiler warnings. The changes applied imply casts that may be hazardous.

+++ source/FormatHandling/clustal_state.cpp
@@ -240,7 +240,7 @@ bool clustal_state::SaveAlignment(const Alignment &alignment, std::ostream *outp
     // Compute maximum sequences name length
     for (i = 0; (i < alignment.originalNumberOfSequences); i++)
         if (alignment.saveSequences[i] != -1)
-            maxLongName = utils::max(maxLongName, alignment.seqsName[i].size());
+            maxLongName = utils::max(maxLongName, (int)alignment.seqsName[i].size());
 
     /* Print alignment header */
     if(!alignment.alignmentInfo.empty() && alignment.alignmentInfo.substr(0,7) == "CLUSTAL")
